added new WalletNodeProvider.tsx for providing wallet node
