+++ app/contexts/WalletContext.tsx
@@ -1,17 +1,10 @@
-import { JellyfishWallet, WalletHdNode, WalletHdNodeProvider } from '@defichain/jellyfish-wallet'
-import { EncryptedProviderData } from '@defichain/jellyfish-wallet-encrypted'
-import { MnemonicProviderData } from '@defichain/jellyfish-wallet-mnemonic'
+import { JellyfishWallet, WalletHdNode } from '@defichain/jellyfish-wallet'
 import { WhaleWalletAccount } from '@defichain/whale-api-wallet'
 import React, { createContext, PropsWithChildren, useContext, useEffect, useMemo, useState } from 'react'
 import { Logging } from '../api'
-import {
-  initJellyfishWallet,
-  MnemonicEncrypted,
-  MnemonicUnprotected,
-  WalletPersistenceData,
-  WalletType
-} from '../api/wallet'
+import { initJellyfishWallet } from '../api/wallet'
 import { useNetworkContext } from './NetworkContext'
+import { useWalletNodeContext } from './WalletNodeProvider'
 import { useWhaleApiClient } from './WhaleContext'
 
 interface WalletContextI {
@@ -35,77 +28,15 @@ export function useWalletContext (): WalletContextI {
   return useContext(WalletContext)
 }
 
-interface WalletProviderProps<T> extends PropsWithChildren<any> {
-  data: WalletPersistenceData<T>
-}
-
-/**
- * Automatically determine the correct WalletProvider to use based on the wallet type.
- */
-export function WalletProvider (props: WalletProviderProps<any>): JSX.Element | null {
-  if (props.data.type === WalletType.MNEMONIC_UNPROTECTED) {
-    return (
-      <MnemonicUnprotectedProvider {...props}>
-        {props.children}
-      </MnemonicUnprotectedProvider>
-    )
-  }
-
-  if (props.data.type === WalletType.MNEMONIC_ENCRYPTED) {
-    return (
-      <MnemonicEncryptedProvider {...props}>
-        {props.children}
-      </MnemonicEncryptedProvider>
-    )
-  }
-
-  throw new Error(`wallet type: ${props.data.type as string} not available`)
-}
-
-function MnemonicUnprotectedProvider (props: WalletProviderProps<MnemonicProviderData>): JSX.Element | null {
-  const { network } = useNetworkContext()
-
-  const provider = useMemo(() => {
-    return MnemonicUnprotected.initProvider(props.data, network)
-  }, [network, props.data])
-
-  return (
-    <WalletContextProvider provider={provider}>
-      {props.children}
-    </WalletContextProvider>
-  )
-}
-
-function MnemonicEncryptedProvider (props: WalletProviderProps<EncryptedProviderData>): JSX.Element | null {
-  const { network } = useNetworkContext()
-
-  const provider = useMemo(() => {
-    return MnemonicEncrypted.initProvider(props.data, network, {
-      prompt: async () => {
-        /**
-         * wallet context only use for READ purpose (non signing)
-         * see {@link TransactionAuthorization} for signing implementation
-         */
-        throw new Error('No UI attached for passphrase prompting')
-      }
-    })
-  }, [network, props.data])
-
-  return (
-    <WalletContextProvider provider={provider}>
-      {props.children}
-    </WalletContextProvider>
-  )
-}
-
-function WalletContextProvider (props: PropsWithChildren<{ provider: WalletHdNodeProvider<WalletHdNode> }>): JSX.Element | null {
+export function WalletContextProvider (props: PropsWithChildren<{}>): JSX.Element | null {
+  const { provider } = useWalletNodeContext()
   const [address, setAddress] = useState<string>()
   const { network } = useNetworkContext()
   const client = useWhaleApiClient()
 
   const wallet = useMemo(() => {
-    return initJellyfishWallet(props.provider, network, client)
-  }, [props.provider, network, client])
+    return initJellyfishWallet(provider, network, client)
+  }, [provider, network, client])
 
   useEffect(() => {
     wallet.get(0).getAddress().then(value => {
