+++ app/screens/RootNavigator.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
-import { WalletProvider } from '../contexts/WalletContext'
+import { WalletContextProvider } from '../contexts/WalletContext'
+import { WalletNodeProvider } from '../contexts/WalletNodeProvider'
 import { useWalletPersistenceContext } from '../contexts/WalletPersistenceContext'
 import { AppNavigator } from './AppNavigator/AppNavigator'
 import { TransactionAuthorization } from './TransactionAuthorization'
@@ -21,9 +22,11 @@ export function RootNavigator (): JSX.Element {
   }
 
   return (
-    <WalletProvider data={wallets[0]}>
-      <TransactionAuthorization />
-      <AppNavigator />
-    </WalletProvider>
+    <WalletNodeProvider data={wallets[0]}>
+      <WalletContextProvider>
+        <TransactionAuthorization />
+        <AppNavigator />
+      </WalletContextProvider>
+    </WalletNodeProvider>
   )
 }
